Feature: Introducing Static Code Analysis, PHPStan and refactor at level 2
