+++ composer.json
@@ -30,8 +30,8 @@
         }
     ],
     "autoload": {
-        "psr-0": {
-            "Storyblok": "src"
+        "psr-4": {
+            "Storyblok\\": "src/Storyblok"
         }
     },
     "autoload-dev": {
@@ -49,13 +49,21 @@
     },
     "scripts": {
         "test": "vendor/bin/pest",
+        "test-ci": "vendor/bin/pest --ci",
         "phpstan": "phpstan analyse",
         "test-coverage": "vendor/bin/pest --coverage",
         "format": "PHP_CS_FIXER_IGNORE_ENV=1 php-cs-fixer fix --allow-risky=yes --using-cache=no",
         "cs": "PHP_CS_FIXER_IGNORE_ENV=1 php-cs-fixer fix --dry-run --using-cache=no",
         "all-check": [
             "@format",
+            "@phpstan",
             "@test"
+        ],
+        "all-check-ci": [
+            "@cs",
+            "@phpstan",
+            "@test-ci"
         ]
+
     }
 }
