+++ phpstan.neon
@@ -1,4 +1,4 @@
 parameters:
-    level: 0
+    level: 2
     paths:
         - src
